Start of API refactor
Move to packages
See APIStructure.md for layout
