+++ src/main/java/gr8pefish/ironbackpacks/api/upgrade/BackpackUpgrade.java
@@ -1,8 +1,11 @@
-package gr8pefish.ironbackpacks.api;
+package gr8pefish.ironbackpacks.api.upgrade;
 
 import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Sets;
+import gr8pefish.ironbackpacks.api.backpack.BackpackInfo;
+import gr8pefish.ironbackpacks.api.IronBackpacksAPI;
+import gr8pefish.ironbackpacks.api.ModifyMethod;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
