+++ src/main/java/gr8pefish/ironbackpacks/api/upgrade/IUpgrade.java
@@ -1,6 +1,8 @@
-package gr8pefish.ironbackpacks.api;
+package gr8pefish.ironbackpacks.api.upgrade;
 
 import com.google.common.base.Preconditions;
+import gr8pefish.ironbackpacks.api.IronBackpacksAPI;
+import gr8pefish.ironbackpacks.api.upgrade.BackpackUpgrade;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
