+++ src/main/java/gr8pefish/ironbackpacks/core/RegistrarIronBackpacks.java
@@ -4,6 +4,11 @@ import com.google.common.base.Preconditions;
 import gr8pefish.ironbackpacks.ConfigHandler;
 import gr8pefish.ironbackpacks.IronBackpacks;
 import gr8pefish.ironbackpacks.api.*;
+import gr8pefish.ironbackpacks.api.backpack.BackpackInfo;
+import gr8pefish.ironbackpacks.api.backpack.IBackpack;
+import gr8pefish.ironbackpacks.api.upgrade.BackpackUpgrade;
+import gr8pefish.ironbackpacks.api.upgrade.IUpgrade;
+import gr8pefish.ironbackpacks.api.variant.BackpackType;
 import gr8pefish.ironbackpacks.item.ItemBackpack;
 import gr8pefish.ironbackpacks.item.ItemUpgrade;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
@@ -48,7 +53,7 @@ public class RegistrarIronBackpacks {
     @GameRegistry.ObjectHolder("diamond")
     public static final BackpackType PACK_DIAMOND = null;
 
-    // Backpack upgrades
+    // Backpack upgrade
     @GameRegistry.ObjectHolder("damage_bar")
     public static final BackpackUpgrade UPGRADE_DAMAGE_BAR = null;
     @GameRegistry.ObjectHolder("lock")
@@ -98,7 +103,7 @@ public class RegistrarIronBackpacks {
             if (stack.getItem() instanceof IBackpack) {
                 IBackpack backpack = (IBackpack) stack.getItem();
                 BackpackInfo backpackInfo = backpack.getBackpackInfo(stack);
-                ResourceLocation location = new ResourceLocation(backpackInfo.getBackpackType().getIdentifier().getResourceDomain(), "backpack/" + backpackInfo.getBackpackType().getIdentifier().getResourcePath());
+                ResourceLocation location = new ResourceLocation(backpackInfo.getBackpackVariant().getIdentifier().getResourceDomain(), "backpack/" + backpackInfo.getBackpackVariant().getIdentifier().getResourcePath());
                 return new ModelResourceLocation(location, "inventory");
             }
             return new ModelResourceLocation(new ResourceLocation(IronBackpacks.MODID, "backpack/null"), "inventory");
@@ -147,7 +152,7 @@ public class RegistrarIronBackpacks {
                 .create();
 
         REGISTRY_UPGRADES = new RegistryBuilder<BackpackUpgrade>()
-                .setName(new ResourceLocation(IronBackpacks.MODID, "upgrades"))
+                .setName(new ResourceLocation(IronBackpacks.MODID, "upgrade"))
                 .setDefaultKey(IronBackpacksAPI.NULL)
                 .setType(BackpackUpgrade.class)
                 .setIDRange(0, Integer.MAX_VALUE - 1)
