+++ src/main/java/gr8pefish/ironbackpacks/item/ItemUpgrade.java
@@ -2,8 +2,8 @@ package gr8pefish.ironbackpacks.item;
 
 import com.google.common.collect.Lists;
 import gr8pefish.ironbackpacks.IronBackpacks;
-import gr8pefish.ironbackpacks.api.BackpackUpgrade;
-import gr8pefish.ironbackpacks.api.IUpgrade;
+import gr8pefish.ironbackpacks.api.upgrade.BackpackUpgrade;
+import gr8pefish.ironbackpacks.api.upgrade.IUpgrade;
 import gr8pefish.ironbackpacks.api.IronBackpacksAPI;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.FontRenderer;
